requirements.txt
fastapi>=0.115.6
uvicorn>=0.32.1
dependency-injector>=4.43.0
sqlalchemy>=2.0.36
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3
pytest-asyncio>=0.15.1
httpx>=0.18.2
aiosqlite>=0.20.0
pydantic>=2.10.3
six>=1.16.0
temporalio>=1.8.0
tenacity>=9.0.0
pydantic-settings>=2.7.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
passlib>=1.7.4
pydantic[email]>=2.10.3
bcrypt>=4.2.1
aiomysql>=0.2.0
aiosqlite>=0.20.0
PyMySQL>=1.1.1
colorlog>=6.9.0
loguru>=0.7.3
orjson>=3.10.12
redis>=5.2.1
langchain>=0.3.14
langchain-core>=0.3.29
langchain-openai>=0.2.13
prometheus_client>=0.21.1
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp>=1.29.0
pytest>=8.3.4
pytest-asyncio>=0.25.0
opentelemetry-instrumentation-fastapi>=0.50b0
celery>=5.4.0
opentelemetry-instrumentation-celery>=0.50b0
opentelemetry-instrumentation-threading>=0.50b0
opentelemetry.instrumentation.system_metrics>=0.50b0
rich>=13.9.4
structlog>=24.4.0
sqlmodel>=0.0.22
alembic>=1.14.0
pymysql>=1.1.1
aiomysql>=0.2.0
langchain_community>=0.3.14
langchain-core>=0.3.28
langchain_experimental>=0.3.4
langchain_neo4j>=0.2.0
termcolor>=2.5.0
python-multipart>=0.0.20
minio>=7.2.14
pymilvus>=2.5.3
opensearch>=0.9.2
opensearch-py>=2.8.0
langchain_anthropic>=0.3.1
langchain_chroma>=0.2.0
langchain_milvus>=0.1.7
tokenizers>=0.21.0
unstructured>=0.16.12
pdfminer>=20191125
langchain_opensearch>=0.0.2
langchain_ollama>=0.2.2
langchain_huggingface>=0.1.2

